LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-container-style-expected.txt


FAIL Style query on #target should query parent element assert_equals: expected "D" but got "E"
FAIL Style query on ::before should query originating element assert_equals: expected "C" but got "E"
